pyproject.toml
[tool.poetry]
name = "jkh"
version = "0.1.0"
description = ""
authors = ["Sandi_Navoiy <santes.au@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.9"
flake8 = "^6.1.0"
django = "^4.2.7"
black = "^23.11.0"
redis = "^5.0.1"
crispy-bootstrap4 = "^2023.1"
pillow = "^10.1.0"
django-phonenumber-field = "^7.3.0"
django-phone-field = "^1.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
